+++ Chatterino/changelog.md
@@ -1,5 +1,10 @@
 ﻿# changelog
 
+## 0.4
+- added login via fourtf.com for users that can't open a tcp port
+- added /mentions tab (thanks to pajlada)
+- tab gets highlighted when you get a whispers
+
 ## 0.3.1.1
 - added ffz event emotes